requirements.txt
# Core dependencies
requests>=2.28.0
yt-dlp>=2023.7.6
pyyaml>=6.0
schedule>=1.2.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
flake8>=6.0.0
black>=23.0.0
mypy>=1.0.0
